+++ src/main/java/com/vbaslak/issuetracker/domain/Status.java
@@ -3,10 +3,7 @@ package com.vbaslak.issuetracker.domain;
 import org.springframework.security.core.GrantedAuthority;
 
 public enum Status implements GrantedAuthority {
-    CREATED,
-    IN_PROGRESS,
-    RESOLVED,
-    CLOSED;
+    CREATED, IN_PROGRESS, RESOLVED, CLOSED;
 
     @Override
     public String getAuthority() {
