+++ src/main/java/com/vbaslak/issuetracker/repos/CommentRepository.java
@@ -1,11 +1,11 @@
-package com.vbaslak.issuetracker.repos;
-
-import com.vbaslak.issuetracker.domain.Comment;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.List;
-
-public interface CommentRepository extends CrudRepository<Comment, Long> {
-    List<Comment> findByComment(String comment);
-}
-
+//package com.vbaslak.issuetracker.repos;
+//
+//import com.vbaslak.issuetracker.domain.Comment;
+//import org.springframework.data.repository.CrudRepository;
+//
+//import java.util.List;
+//
+//public interface CommentRepository extends CrudRepository<Comment, Long> {
+//    List<Comment> findByComment(String comment);
+//}
+//
