+++ src/main/resources/templates/greeting.ftl
@@ -1,5 +1,5 @@
 <#import "parts/common.ftl" as c>
 <@c.page>
 <h1>Пользователь, добрый день! Добро пожаловать на Issue Tracker</h1>
-<a href="/main"><h2>Перейти на главную страницу</h2></a>
+<a href="/main"><h2>Перейти к списку задач</h2></a>
 </@c.page>
